refactor: separate entrypoint file to utils & workflow to workflowHelper file & add types
