+++ tools/spec-gen-sdk/.prettierrc
@@ -1,0 +1,8 @@
+{
+    "semi": true,
+    "trailingComma": "all",
+    "singleQuote": true,
+    "tabWidth": 2,
+    "printWidth": 180,
+    "endOfLine": "auto"
+}
