+++ tools/spec-gen-sdk/src/utils/fsUtils.ts
@@ -1,8 +1,8 @@
 import path from 'path';
 import fs from 'fs';
 import { any } from './arrays';
-import { WorkflowContext } from '../automation/workflow';
 import { toolError } from './messageUtils';
+import { WorkflowContext } from '../types/Workflow';
 
 export const writeTmpJsonFile = (context: WorkflowContext, fileName: string, content: unknown) => {
   const filePath = path.join(context.tmpFolder, fileName);
